+++ app/controllers/rpi_auth/application_controller.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module RpiAuth
-  class ApplicationController < ActionController::Base
-  end
-end