+++ app/controllers/rpi_auth/auth_controller.rb
@@ -1,27 +1,24 @@
 # frozen_string_literal: true
 
+require 'rpi_auth/controllers/current_user'
+
 module RpiAuth
-  class AuthController < ApplicationController
+  class AuthController < ActionController::Base
+    include RpiAuth::Controllers::CurrentUser
+
     protect_from_forgery with: :null_session
 
-    # rubocop:disable Metrics/AbcSize
     def callback
       # Prevent session fixation. If the session has been initialized before
       # this, and we need to keep the data, then we should copy values over.
       reset_session
 
       auth = request.env['omniauth.auth']
-      user = RpiAuth.user_model.from_omniauth(auth)
-      session[:current_user] = user
+      self.current_user = RpiAuth.user_model.from_omniauth(auth)
 
-      return redirect_to RpiAuth.configuration.success_redirect if RpiAuth.configuration.success_redirect
-
-      if request.env.key?('omniauth.origin') && request.env['omniauth.origin'].present?
-        return redirect_to(request.env['omniauth.origin'], allow_other_host: false)
-      end
-
-      redirect_to '/'
-      # rubocop:enable Metrics/AbcSize
+      redirect_to RpiAuth.configuration.success_redirect.presence ||
+                  request.env.fetch('omniauth.origin', nil).presence ||
+                  '/'
     end
 
     def destroy